moved each provider to its own spider
